refactor(models): replace user interfaces with types
