--- src/app/core/models/enums/roles.enum.ts
@@ -1,5 +1,0 @@
-export enum Role {
-  UNAUTHORIZED = 'UNAUTHORIZED',
-  ADMIN = 'ADMIN',
-  USER = 'USER',
-}
