--- src/app/core/models/user-response.model.ts
@@ -1,6 +1,0 @@
-export interface UserResponse {
-  id: string;
-  username: string;
-  email: string;
-  roles: string[];
-}
