--- src/app/core/models/user.model.ts
@@ -1,8 +1,0 @@
-import { Role } from './enums/roles.enum';
-
-export interface User {
-  id: number;
-  username: string;
-  email: string;
-  roles: Role[];
-}
