+++ src/app/core/models/user.type.ts
@@ -1,0 +1,13 @@
+import { Role } from './role.type';
+
+export type UserResponse = {
+  id: string;
+  username: string;
+  email: string;
+  roles: string[];
+};
+
+export type User = Omit<UserResponse, 'id' | 'roles'> & {
+  id: number;
+  roles: Role[];
+};
