+++ src/app/core/services/user.service.ts
@@ -1,7 +1,7 @@
 import { BehaviorSubject, Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
 
-import { User } from '../models/user.model';
+import { User } from '../models/user.type';
 
 @Injectable({
   providedIn: 'root',
